coverart:script now respects --vid and --audio-display options
Also made several changes to how the script operates.
Audio files are now found by checking the first track in the track list.

If audio-display is disabled, the script will not add any tracks, not
even the placeholder.

If vid=0 is chosen, then all the coverart is loaded normally, but it is
not selected be default.
If a different vid value is selected, then the script will actually
start with that number cover art selected. I'm not exactly sure why this
works, and it may not work in all situations.

There does seem to be some weird behaviour when setting a vid of
greater than 1, while load_extra_tracks is disabled. I believe this is
due to some weird behaviour in the mpv api when setting vid values
greater than the actual number of vid tracks.

Regardless, setting a vid of greater than 1 when load_extra_tracks is
enabled makes no sense, so I'm just ignoring this bug.