placeholder: fixed bug for files with embedded tracks
If a file that required a placeholder was played directly before a
file with embedded art, then the placeholder was automatically
being loaded due to script's caching of the same directory.

To fix this, simply don't save the placeholder in the cache and run the
usual placeholder checks if the cover cache is empty.